+++ src/main/java/io/xlorey/FluxLoader/utils/BackupTools.java
@@ -1,7 +1,9 @@
 package io.xlorey.FluxLoader.utils;
 
 import lombok.experimental.UtilityClass;
+import org.objectweb.asm.*;
 
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -82,6 +84,36 @@ public class BackupTools {
     }
 
     /**
+     * Checks for the presence of the Modified annotation in the class file's methods.
+     * This method reads a class file specified by the given path and analyzes it to
+     * determine if any of its methods are annotated with the Modified annotation.
+     * If such an annotation is found, it logs a warning and throws an IllegalStateException.
+     * @param classFilePath The path to the class file to be checked.
+     * @throws IOException If an I/O error occurs during reading the class file.
+     */
+    private static void checkAnnotationModified(Path classFilePath) throws IOException {
+        try (FileInputStream fileInputStream = new FileInputStream(classFilePath.toFile())) {
+            ClassReader reader = new ClassReader(fileInputStream);
+
+            reader.accept(new ClassVisitor(Opcodes.ASM9) {
+                @Override
+                public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
+                    return new MethodVisitor(Opcodes.ASM9) {
+                        @Override
+                        public AnnotationVisitor visitAnnotation(String descriptor, boolean visible) {
+                            if (descriptor.equals("Lio/xlorey/FluxLoader/annotations/Modified;")) {
+                                Logger.printSystem("Signs of modification detected! Please, before installing a new version, remove the old one using the '--uninstall' flag!");
+                                throw new IllegalStateException(String.format("Annotation 'Modified' found in method '%s' of file '%s'", name, classFilePath.toAbsolutePath()));
+                            }
+                            return super.visitAnnotation(descriptor, visible);
+                        }
+                    };
+                }
+            }, 0);
+        }
+    }
+
+    /**
      * Creating a backup file
      * @param pathToClassFile path to the file with '.class' extension
      * @throws IOException in cases of unsuccessful backup creation
@@ -92,6 +124,9 @@ public class BackupTools {
         validateClassFilePath(pathToClassFile);
 
         Path originalFilePath = getFullPath(pathToClassFile);
+
+        checkAnnotationModified(originalFilePath);
+
         Path backupFilePath = getBackupPath(originalFilePath);
 
         Path directoryPath = originalFilePath.getParent();